Fixx exceptoin

+++ heobs_image_classification.py
@@ -27,22 +27,6 @@ def heobs_image_classification(template, max_iter, img_width, img_height, gpu_id
         logger.debug("Working dir: %s", ws.workspace(""))
         os.chdir(os.path.dirname(os.path.realpath(__file__)))
         classes = ["being", "heritage", "scenery"]
-        train_zip = GoogleFile('0BzL8pCLanAIAd0hBV2NUVHpmckE', ws.workspace('data/heobs_large_dataset.zip'))
-
-        pycaffe = PyCaffe()
-        queue.put(("update", test_id, 2, 100, "downloading dataset..."))
-        google_download = DownloadGoogleDrive()
-
-        logger.debug("\n\n------------------------PREPARE PHRASE----------------------------\n\n")
-
-        logger.debug("Starting download train file")
-        google_download.download_file_from_google_drive(train_zip)
-        logger.debug("Finish")
-
-        logger.debug("Extracting train zip file")
-        queue.put(("update", test_id, 10, 100, "extracting dataset..."))
-        unzip_with_progress(train_zip.file_path, ws.workspace("data/extracted"))
-        logger.debug("Finish")
 
         train_lmdb_path = ws.workspace("data/extracted/train_lmdb")
 
@@ -62,34 +46,55 @@ def heobs_image_classification(template, max_iter, img_width, img_height, gpu_id
 
         snapshot_prefix = ws.workspace("caffe_model/snapshot")
 
-        queue.put(("update", test_id, 15, 100, "creating lmdb..."))
-        lmdb = CreateLmdb()
-        lmdb.create_lmdb(ws.workspace("data/extracted/heobs_large_dataset"), train_lmdb_path, validation_lmdb_path, classes,
-                         test_path, img_width, img_height)
+        pycaffe = PyCaffe()
 
-        mean_proto = ws.workspace("data/mean.binaryproto")
+        if os.path.isfile(caffe_log):
+            queue.put(("update", test_id, 2, 100, "resuming"))
+        else:
+            train_zip = GoogleFile('0BzL8pCLanAIAd0hBV2NUVHpmckE', ws.workspace('data/heobs_large_dataset.zip'))
 
-        queue.put(("update", test_id, 20, 100, "computing train image mean..."))
-        pycaffe.compute_image_mean("lmdb", train_lmdb_path, mean_proto, logger)
-        queue.put(("update", test_id, 25, 100, "computing test image mean..."))
-        pycaffe.compute_image_mean("lmdb", validation_lmdb_path, mean_proto, logger)
+            queue.put(("update", test_id, 2, 100, "downloading dataset..."))
+            google_download = DownloadGoogleDrive()
 
-        solver_mode = constant.CAFFE_SOLVER
-        if "CAFFE_SOLVER" in os.environ:
-            solver_mode = os.environ['CAFFE_SOLVER']
+            logger.debug("\n\n------------------------PREPARE PHRASE----------------------------\n\n")
 
-        py_render_template("template/" + template + "/caffenet_train.template", caffe_train_model, mean_file=mean_proto,
-                           train_lmdb=train_lmdb_path, validation_lmdb=validation_lmdb_path,
-                           batchsize_train=batchsize_train,
-                           batchsize_test=batchsize_test,
-                           num_output=len(classes))
-        py_render_template("template/" + template + "/caffenet_solver.template", caffe_solver,
-                           caffe_train_model=caffe_train_model,
-                           max_iter=max_iter,
-                           snapshot_prefix=snapshot_prefix,
-                           learning_rate=lr,
-                           stepsize=stepsize,
-                           solver_mode=solver_mode)
+            logger.debug("Starting download train file")
+            google_download.download_file_from_google_drive(train_zip)
+            logger.debug("Finish")
+
+            logger.debug("Extracting train zip file")
+            queue.put(("update", test_id, 10, 100, "extracting dataset..."))
+            unzip_with_progress(train_zip.file_path, ws.workspace("data/extracted"))
+            logger.debug("Finish")
+
+            queue.put(("update", test_id, 15, 100, "creating lmdb..."))
+            lmdb = CreateLmdb()
+            lmdb.create_lmdb(ws.workspace("data/extracted/heobs_large_dataset"), train_lmdb_path, validation_lmdb_path, classes,
+                             test_path, img_width, img_height)
+
+            mean_proto = ws.workspace("data/mean.binaryproto")
+
+            queue.put(("update", test_id, 20, 100, "computing train image mean..."))
+            pycaffe.compute_image_mean("lmdb", train_lmdb_path, mean_proto, logger)
+            queue.put(("update", test_id, 25, 100, "computing test image mean..."))
+            pycaffe.compute_image_mean("lmdb", validation_lmdb_path, mean_proto, logger)
+
+            solver_mode = constant.CAFFE_SOLVER
+            if "CAFFE_SOLVER" in os.environ:
+                solver_mode = os.environ['CAFFE_SOLVER']
+
+            py_render_template("template/" + template + "/caffenet_train.template", caffe_train_model, mean_file=mean_proto,
+                               train_lmdb=train_lmdb_path, validation_lmdb=validation_lmdb_path,
+                               batchsize_train=batchsize_train,
+                               batchsize_test=batchsize_test,
+                               num_output=len(classes))
+            py_render_template("template/" + template + "/caffenet_solver.template", caffe_solver,
+                               caffe_train_model=caffe_train_model,
+                               max_iter=max_iter,
+                               snapshot_prefix=snapshot_prefix,
+                               learning_rate=lr,
+                               stepsize=stepsize,
+                               solver_mode=solver_mode)
 
         logger.debug("\n\n------------------------TRAINING PHRASE-----------------------------\n\n")
 
@@ -103,41 +108,42 @@ def heobs_image_classification(template, max_iter, img_width, img_height, gpu_id
 
         logger.debug("\n\n------------------------TESTING PHRASE-----------------------------\n\n")
 
-        queue.put(("update", test_id, 90, 100, "starting to test..."))
-        py_render_template("template/" + template + "/caffenet_deploy.template", caffe_deploy,
-                           num_output=len(classes), img_width=img_width, img_height=img_height)
+        if not os.path.isfile(ws.workspace("result/slover.prototxt")):
+            queue.put(("update", test_id, 90, 100, "starting to test..."))
+            py_render_template("template/" + template + "/caffenet_deploy.template", caffe_deploy,
+                               num_output=len(classes), img_width=img_width, img_height=img_height)
 
-        with Silence(stderr=ws.workspace("tmp/output.txt"), mode='w'):
-            set_caffe_gpu(gpu_id)
-            logger.debug("\nReading mean file")
-            queue.put(("update", test_id, 91, 100, "reading mean file..."))
-            mean_data = read_mean_data(mean_proto)
+            with Silence(stderr=ws.workspace("tmp/output.txt"), mode='w'):
+                set_caffe_gpu(gpu_id)
+                logger.debug("\nReading mean file")
+                queue.put(("update", test_id, 91, 100, "reading mean file..."))
+                mean_data = read_mean_data(mean_proto)
 
-            logger.debug("\nReading neural network model")
-            queue.put(("update", test_id, 92, 100, "reading cnn model..."))
-            net = read_model_and_weight(caffe_deploy, snapshot_prefix + "_iter_" + str(max_iter) + ".caffemodel")
-            transformer = image_transformers(net, mean_data)
+                logger.debug("\nReading neural network model")
+                queue.put(("update", test_id, 92, 100, "reading cnn model..."))
+                net = read_model_and_weight(caffe_deploy, snapshot_prefix + "_iter_" + str(max_iter) + ".caffemodel")
+                transformer = image_transformers(net, mean_data)
 
-            logger.debug("Predicting...")
-            queue.put(("update", test_id, 95, 100, "predicting..."))
-            prediction = making_predictions(ws.workspace("data/extracted/test"), transformer, net, img_width, img_height)
+                logger.debug("Predicting...")
+                queue.put(("update", test_id, 95, 100, "predicting..."))
+                prediction = making_predictions(ws.workspace("data/extracted/test"), transformer, net, img_width, img_height)
 
-        with open(ws.workspace("tmp/output.txt"), 'r') as f:
-            logger.debug(f.read())
+            with open(ws.workspace("tmp/output.txt"), 'r') as f:
+                logger.debug(f.read())
 
-        queue.put(("update", test_id, 100, 100, "exporting data..."))
-        logger.debug("Exporting result to csv")
-        export_to_csv(prediction, ws.workspace("result/test_result.csv"))
+            queue.put(("update", test_id, 100, 100, "exporting data..."))
+            logger.debug("Exporting result to csv")
+            export_to_csv(prediction, ws.workspace("result/test_result.csv"))
 
-        logger.debug("Exporting predict result to folder")
-        export_data(prediction, ws.workspace("data/extracted/test"), ws.workspace("result/data"))
+            logger.debug("Exporting predict result to folder")
+            export_data(prediction, ws.workspace("data/extracted/test"), ws.workspace("result/data"))
 
-        logger.debug("Moving log")
-        shutil.copyfile(caffe_log, ws.workspace("result/caffe_train.log"))
-        shutil.copyfile(caffe_train_model, ws.workspace("result/model/train_val.prototxt"))
-        shutil.copyfile(caffe_deploy, ws.workspace("result/deploy.prototxt"))
-        shutil.copyfile(caffe_solver, ws.workspace("result/slover.prototxt"))
-        logger.debug("\n\n-------------------------FINISH------------------------------------\n\n")
+            logger.debug("Moving log")
+            shutil.copyfile(caffe_log, ws.workspace("result/caffe_train.log"))
+            shutil.copyfile(caffe_train_model, ws.workspace("result/model/train_val.prototxt"))
+            shutil.copyfile(caffe_deploy, ws.workspace("result/deploy.prototxt"))
+            shutil.copyfile(caffe_solver, ws.workspace("result/slover.prototxt"))
+            logger.debug("\n\n-------------------------FINISH------------------------------------\n\n")
 
         logger.debug("\nTest completed")
         queue.put(("done", test_id, "completed"))
